Moves the ivPopupButton
